Movement updated, multiple bugs addressed, UML created
I went through the movement and tightened up some of the functionality,
created some functions for re-used codes, stream lined it and fixed any
found bugs.

There were still plenty of bugs in the GUI, went through those for
fixes. Only the current player can press any buttons and it is onlt the
roll and their pieces they can press. Disabled the endTurn during other
players turn and the playTrouble once a game has started. Added names to
any messages sent so that you know who is chatting in the interface.
Also says whose turn it is.

Lastly downloaded addion for Java that is an easyUML Diagramer. Added
that as another project to the Trouble game. Selected GUI, Echo Server,
Chat Client, Trouble, Player, and Piece classes to diagram and created
them.

*** Still one current bug that I can't seem to fix. When a 6 is rolled
if it is the Red player then they do not get another turn. The
application moves to the ext player. Makes no sense seeing as when a six
is rolled and the other players select the piece for movement the game
auto updates and they are ready to roll again without having to endTurn.
For red it acts like a regular roll and the board doesn't update.